in progress with refactoring still. provider_id logic requires testing
